Rename files and add SciPy to requirements.txt
