+++ requirements.txt
@@ -3,6 +3,6 @@ diskcache==5.6.3
 dash[diskcache]
 dwave-ocean-sdk>=4.1
 matplotlib>=3.3.4
-mip>=1.15.0
+scipy>=1.10.0
 pandas==2.1.3
 tabulate>=0.8.9
